evaluate the cast operator client side.
I incorporated this because I needed to explicitly cast something from
the DB as a date or interval and I needed to be able to test whether
an object was a member of some generated query. Alas, doing that
depends (as implemented) on the db.date and db.time objects which I
use from the algebraic-data-types package. I haven't published those
as open source yet, so I've #+...ed those bits out from here.